Code clean up to remove warning about implicit cast. (Cast is required since NSData -length returns an unsinged integer and OpenSSL length parameters work with a signed integer)

+++ MIHCrypto/AES/MIHAESKey.m
@@ -197,7 +197,7 @@
         return nil;
     }
 
-    if (!EVP_EncryptUpdate(&encryptCtx, cipherBytes, (int *) &blockLength, messageData.bytes, messageData.length)) {
+    if (!EVP_EncryptUpdate(&encryptCtx, cipherBytes, (int *) &blockLength, messageData.bytes, (int)messageData.length)) {
         if (error) *error = [NSError errorFromOpenSSL];
         return nil;
     }
